Changed LLVM from building release+symbols to just release.
Having symbols is not super necessary and causes the build folder to take 5x as much space and the resulting binaries to be ~2x as large.
We don't have a great way to distribute the symbols in a package either since NuGet symbol servers don't work with full PDBs.
I may experiment with distributing a fatter ClangSharp.Pathogen.Runtime-Symbols package in the future, but for now I won't bother.