Changed ValidateAndAppendBuildMetadataToVersionNumber to be skipped for design-time builds.
The real cause of the problems fixed in 1f903ee6afa2d4f2715d83aa442217285d38f27c was Git information from SourceLink not being available during design-time builds.
It's not really important that the build metadata is present in design-time builds, so there's no practical reason to run this target.
I decided to leave using InformationalVersion instead of Version because realistically the build metdata shouldn't be part of the non-informational version anyway since it is stripped later on.